+++ Models/Server.Models/DigitalPasses/LoyaltyCards/Apple_Passes_Loyalty_Card_Model.cs
@@ -8,9 +8,9 @@ namespace Server.Models
 {
     public class Apple_Passes_Loyalty_Card_Model:WalletPassModel
     {
-        public string Program_Name       { get; set; }
-        public int    Points_Balance     { get; set; }
-        public string Reward_Details     { get; set; }
+        public string  Program_Name       { get; set; }
+        public decimal Points_Balance     { get; set; }
+        public string  Reward_Details     { get; set; }
     }
 
 }
